requirements.txt
pylint==2.4.4
requests==2.22.0
ipython==7.12.0
ipdb==0.12.3
python-dotenv=0.11.0